Remove layout nav.html.heex file
This file is not used as the navigation html code is defined
in root.html.heex.

We could have transform this file to use it as a Phoenix component,
however because the nav is only used once I didn't think it makes
sense to have a reusable component for the nav.